Move `ParamRange` to the `subgrid` module
