+++ neopdf/src/gridpdf.rs
@@ -24,7 +24,7 @@ use super::interpolator::{DynInterpolator, InterpolatorFactory};
 use super::metadata::MetaData;
 use super::parser::SubgridData;
 use super::strategy::AlphaSCubicInterpolation;
-use super::subgrid::{RangeParameters, SubGrid};
+use super::subgrid::{ParamRange, RangeParameters, SubGrid};
 
 /// Errors that can occur during PDF grid operations.
 #[derive(Debug, Error)]
@@ -40,40 +40,6 @@ pub enum Error {
     /// Error indicating invalid interpolation parameters, with a descriptive message.
     #[error("Invalid interpolation parameters: {0}")]
     InterpolationError(String),
-}
-
-/// Represents the valid range of a parameter, with a minimum and maximum value.
-#[derive(Debug, Clone, Copy, Deserialize, Serialize)]
-pub struct ParamRange {
-    /// The minimum value of the parameter.
-    pub min: f64,
-    /// The maximum value of the parameter.
-    pub max: f64,
-}
-
-impl ParamRange {
-    /// Creates a new `ParamRange`.
-    ///
-    /// # Arguments
-    ///
-    /// * `min` - The minimum value.
-    /// * `max` - The maximum value.
-    pub fn new(min: f64, max: f64) -> Self {
-        Self { min, max }
-    }
-
-    /// Checks if a given value is within the parameter range (inclusive).
-    ///
-    /// # Arguments
-    ///
-    /// * `value` - The value to check.
-    ///
-    /// # Returns
-    ///
-    /// `true` if the value is within the range, `false` otherwise.
-    pub fn contains(&self, value: f64) -> bool {
-        value >= self.min && value <= self.max
-    }
 }
 
 /// Stores the complete PDF grid data, including all subgrids and flavor information.
@@ -364,13 +330,6 @@ impl GridPDF {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_param_range() {
-        let range = ParamRange::new(1.0, 10.0);
-        assert!(range.contains(5.0));
-        assert!(!range.contains(15.0));
-    }
 
     #[test]
     fn test_grid_array_creation() {
